"11:05am.
```
The learning rate is 2 ** -11.5
At iteration 9 the cost and the sign accuracy are 25.54138 and 0.804/1.
At iteration 19 the cost and the sign accuracy are 24.15172 and 0.884/1.
At iteration 29 the cost and the sign accuracy are 21.46856 and 0.97/1.
At iteration 39 the cost and the sign accuracy are 20.74911 and 0.97/1.
At iteration 49 the cost and the sign accuracy are 17.63033 and 0.964/1.
At iteration 59 the cost and the sign accuracy are 18.00328 and 0.974/1.
At iteration 69 the cost and the sign accuracy are 15.86308 and 0.978/1.
At iteration 79 the cost and the sign accuracy are 11.49514 and 0.992/1.
At iteration 89 the cost and the sign accuracy are 11.13326 and 0.994/1.
At iteration 99 the cost and the sign accuracy are 10.85187 and 0.992/1.
At iteration 109 the cost and the sign accuracy are 8.973773 and 0.988/1.
At iteration 119 the cost and the sign accuracy are 5.970519 and 1/1.
At iteration 129 the cost and the sign accuracy are 7.688843 and 0.992/1.
At iteration 139 the cost and the sign accuracy are 9.302932 and 0.984/1.
At iteration 149 the cost and the sign accuracy are 8.146052 and 0.994/1.
At iteration 159 the cost and the sign accuracy are 6.196711 and 0.992/1.
At iteration 169 the cost and the sign accuracy are 5.657013 and 0.99/1.
At iteration 179 the cost and the sign accuracy are 6.672916 and 0.986/1.
At iteration 189 the cost and the sign accuracy are 5.669213 and 0.992/1.
At iteration 199 the cost and the sign accuracy are 4.204844 and 0.994/1.
The time was 00:00:10.1028363 for: Training
```

This is it.  Let me compare it with the other one.

```
The learning rate is 2 ** -11.5
At iteration 9 the cost and the sign accuracy are 25.63614 and 0.802/1.
At iteration 19 the cost and the sign accuracy are 24.15467 and 0.9/1.
At iteration 29 the cost and the sign accuracy are 21.70532 and 0.97/1.
At iteration 39 the cost and the sign accuracy are 20.69831 and 0.974/1.
At iteration 49 the cost and the sign accuracy are 18.0439 and 0.968/1.
At iteration 59 the cost and the sign accuracy are 18.0083 and 0.98/1.
At iteration 69 the cost and the sign accuracy are 15.94441 and 0.986/1.
At iteration 79 the cost and the sign accuracy are 12.43328 and 0.992/1.
At iteration 89 the cost and the sign accuracy are 11.221 and 0.998/1.
At iteration 99 the cost and the sign accuracy are 11.11839 and 0.992/1.
At iteration 109 the cost and the sign accuracy are 9.006557 and 0.994/1.
At iteration 119 the cost and the sign accuracy are 6.170288 and 1/1.
At iteration 129 the cost and the sign accuracy are 7.625436 and 0.992/1.
At iteration 139 the cost and the sign accuracy are 8.560534 and 0.988/1.
At iteration 149 the cost and the sign accuracy are 7.026469 and 0.998/1.
At iteration 159 the cost and the sign accuracy are 6.090001 and 0.992/1.
At iteration 169 the cost and the sign accuracy are 5.278853 and 0.994/1.
At iteration 179 the cost and the sign accuracy are 5.739595 and 0.99/1.
At iteration 189 the cost and the sign accuracy are 5.366251 and 0.992/1.
At iteration 199 the cost and the sign accuracy are 4.160935 and 0.986/1.
The time was 00:00:31.2869304 for: Training
```

Ah...

I think I won't make an issue of these differences. They are probably due to numerical instability.

Good work me. Let me time the two. I'll append the timings to the two above when it is done.

11:20am. Wow, fully 3x faster. Yeah, it can definitely be felt. Rather than KFAC it, what I am going to do now is something I've been meaning to do for a while. Add the `__restrict__` annotations for the global arrays in the Cuda kernel. That could potentially make things faster. I am interested in whether it will make any kind of difference.

11:25am. I'll do that after the break. It is time for a well earned rest."