"1:15pm. https://psychobarcodetranslations.wordpress.com/flamering/chapter-10/
Let me stop here. It is time to do some work. I'll add the type annotations in the Cuda code generator and in addition to that watch the Shen talk. I'll probably have time to KFAC the plastic RNN when all of that is done. The error I had yesterday was an exceptional one."