https://devblogs.nvidia.com/cuda-pro-tip-optimize-pointer-aliasing/
Done converting the Spiral compiler. I've put the annotation on the wrong side of the pointer it seems, but I'll test to see if it works like this anyway. Also unfortunately I really cannot put the `const` annotations. Shouldn't the compiler be capable of doing this on its own?

Well, let me give it a try.