"4:20pm.
```
The learning rate is 2 ** -12.5
At iteration 9 the cost and the sign accuracy are 25.9284 and 0.79/1.
At iteration 19 the cost and the sign accuracy are 23.47174 and 0.908/1.
At iteration 29 the cost and the sign accuracy are 18.83529 and 0.976/1.
At iteration 39 the cost and the sign accuracy are 13.49367 and 0.994/1.
At iteration 49 the cost and the sign accuracy are 10.0951 and 0.978/1.
At iteration 59 the cost and the sign accuracy are 8.038506 and 0.99/1.
At iteration 69 the cost and the sign accuracy are 6.308923 and 0.978/1.
At iteration 79 the cost and the sign accuracy are 4.153198 and 0.99/1.
At iteration 89 the cost and the sign accuracy are 3.108134 and 0.994/1.
At iteration 99 the cost and the sign accuracy are 3.450883 and 0.992/1.
At iteration 109 the cost and the sign accuracy are 3.376967 and 0.988/1.
At iteration 119 the cost and the sign accuracy are 1.889218 and 0.994/1.
At iteration 129 the cost and the sign accuracy are 2.496047 and 0.992/1.
At iteration 139 the cost and the sign accuracy are 3.462413 and 0.988/1.
At iteration 149 the cost and the sign accuracy are 3.161825 and 0.986/1.
At iteration 159 the cost and the sign accuracy are 2.980001 and 0.988/1.
At iteration 169 the cost and the sign accuracy are 4.222834 and 0.978/1.
At iteration 179 the cost and the sign accuracy are 3.655708 and 0.984/1.
At iteration 189 the cost and the sign accuracy are 3.133637 and 0.986/1.
At iteration 199 the cost and the sign accuracy are 4.782403 and 0.974/1.
The time was 00:00:22.2677818 for: Training
```

Ugh. The standard plastic_rnn reaps no benefit from this.

4:25pm. Let me try turning it off. I need to fetch the expand_singular from a previous commit."