Tests reveal that waiting on streams make it 5% slower. It is good that I've modified the `expand_singular` to make them optional.
With this I am really done for the day. There is no much point in pushing myself past here.

I am not too tired at this point, but this is no time to starting anything new.

Tomorrow I will give the alternative modulation a try once more. I will have an easier time experimenting now that I do not have to make the backwards pass by hand.

What I will do after that is dust off LN and probably redo it after that so that issues backprop inside the seq kernel. After that I will use it for weight norm.

Though...I had intended to fuse the plastic weights with the regular ones so as to do the multiply in one blast. LN cannot be combined with that as the weights would need to be transposed.

...But for the reverse pass, I would also need to store the intermediate means and standard deviations. I think should use `redo_map` for this sort of thing. It is inevitable - weight norm will make things more complicated, but I'll do it.

...I'll do it. Fusing the weights will definitely be worth it. Forget the LN.

And now that I am reasoning it out, calculating mean and standard deviation in order to backprop for them will really be trivial. The reason why LN was so hard is because I could not split into pieces elegantly. I could do that now, but I am not interested in it anymore.

Let me wind off here.

Everything is coming together nicely. That bug I had yesterday and fixed this morning was just a bump and nothing more.

In less than a week, I will be back in full force and doing RL. If you want to do something, do it with absolute control and pride. So what if I took a little time to do it properly. The time to make a language was exactly in this time period while the RL algorithms are weak. In the 20s, they will be much better. At that point I would not longer have the leisure to sit back and do it properly as sparks will be flying.

When the next generation of neural network processors comes along I will be one of the rare people capable of pushing them to their limit. I still have a while to go, but I am getting there.

I will attain my own power and grasp hold of my talent.