``` The learning rate is 2 ** -12.00083 Starting timing for: iteration 0 The time was 00:00:06.7539986 for: iteration 0 Training: 2.02372045156217 The learning rate is 2 ** -12.00083 Starting timing for: iteration 1 The time was 00:00:06.6897641 for: iteration 1 Training: 1.69079768792932 The learning rate is 2 ** -12.00083 Starting timing for: iteration 2 The time was 00:00:06.7405257 for: iteration 2 Training: 1.63641439120093 The learning rate is 2 ** -12.00083 Starting timing for: iteration 3 The time was 00:00:06.6693839 for: iteration 3 Training: 1.60978936627233 The learning rate is 2 ** -12.00083 Starting timing for: iteration 4 The time was 00:00:06.7751939 for: iteration 4 Training: 1.59044299481491 ```
Here is me taking the covariance calculations out of the `mi'` layer. There is a very marginal improvement. So the story was not that there is no effect when naturalizing fused biases. It is literally the case that naturalizing the biases does nothing. Holy shit.

I cannot believe this.

Let me print out the covariance for back. There has to be something wrong with `expand_singular`.