"10:50am. I am just so blown away by this that I can hardly accept it. Naturalizing the biases is useless apparently.
11:05am. I am still blow away by it.

But in a sense, this is good as the updating the biases is a major overhead that I should be glad to be rid off. On the other hand, I expected that naturalizing the biases would be straightforward extension of natural gradient methods. It seems I was wrong.

11:10am. This morning was such a waste. I like it when I am focused on the goal. Right now I just wanted to do a few tests and test out the alternative modulation scheme, but I got completely sidetracked by this.

11:15am. This is the worst. Nothing is going according to plan and now that it is time to take a break I do not feel like it unlike before.

Forget the alternative modulation scheme for now.

I will take a break and when I come back, dive into numerical analysis. I'll skim the book for an hour or two and then get back into programming. I'll forget this whole bias stuff - I'll apply naturalization to the weights only from here on out which will simplify things significantly.

Without a doubt, the future will hold more sensible optimization methods that will give me true understanding and meshing with the models they are used on. Fuck backpropagation.

For now I will just go with what I have."