I'll admit, despite spending all this time reading the book, I barely read more than a few pages properly.
I've been thinking all this time about RNNs.

Somehow I feel I am starting to come to a fundamental change of how layers should be viewed, and that includes RNNs.

There is a bunch of stuff that I want to try regarding recurrent connections. If it works, I won't need MI for the plastic connection and I will get the modulation to work at all levels.

All this time I've been thinking about the LSTM as a less efficient MI-RNN, but I am starting to think that the LSTM could be viewed as a MI-RNN with the biases dependent on the data.

Also before I settle on the plastic RNN architecture there is something really, really important I need to try - merging the input and state vectors and adding the bias to them. This is something that I am going to be doing for both the static and plastic connections, and when I add reversible learning, the plastic connections are going to be a part of the same matrix.

Honestly, I am at such an advantaged stage here. I am going to completely master the basic RNN at this very point. Flexible tensor tree views are absolutely an indispensable part of that.

Humans are always limited by the tools they use. Now that I have this my perception is shifting once more.

There is one thing I am sure. I should put all the matrix multiplies into a single matrix and apply KFAC to that. That will resolve everything. Today's bias nonsense will not be repeated again.

What I will do tomorrow is going through the RNN designs from the top. Well, I'll leave the `mi'` and `mi''` as they are, but I will redesign the LSTM. I will try taking the recent LSTM ideas and applying them to plastic RNN after that.

The goal has changed - I want to try finding a better plastic RNN architecture that does not use MI-RNN like biases.