https://psychobarcodetranslations.wordpress.com/flamering/chapter-29/
I am still reading the 'The Girl Who Bore the Flame Ring'. After how yesterday turned out, I just do not feel like programming. But I must. Let me stop here and I will do at least something. This is my punishment for not winning the way I wanted. I must prove my commitment to my new path again and again.

It is bothersome and annoying, but let me get into it just for a bit. I've thought about it for quite a while yesterday. For now, let me start work on the perfected LSTM.

There are several ideas I need to implement in order to accomplish it.

For starters I will remove the biases as I do not have a `concat` function capable of acting as a map.

Then as the first step, I will fuse the input and state into the same multiply. In essence I am going a step beyond what was shown in NVidia blog post. This will make the LSTM significantly faster, even beyond that of the much smaller MIRNN.

Besides this, the main thrust is to centralize all the matrix multiplies (and in future convolutions) into a single operations which will leave the rest to a map operation.

This will apply not just to the LSTM, but to all future layers. What I am proposing here is that this vision is in the fact the perfect, idealized cell template.

There is absolutely no need for there to be multiple different matrix multiplies within a single layer and that will include even the plastic RNNs.

Now that I think about it more, the first step should not be the LSTM, but a vanilla RNN. I should just focus on it for now. I'll implement my idea, KFAC it which should be super easy given that that it will be needed only in one place for all future architectures (and does not work on the biases directly anyway). Then I will begin work on a more advanced concat function. Then I will implement the optimization that fills the input into the state vector.

It is a miracle - I am only placing restrictions on what should be done, but rather that holding me back it seems to be releasing vast amounts of power. Constraints when done properly draw out the truth of reality that lies beyond the shadows.