``` The learning rate is 2 ** -12.49694 Starting timing for: iteration 0 The time was 00:00:06.5981770 for: iteration 0 Training: 2.25178796120728 The learning rate is 2 ** -12.49694 Starting timing for: iteration 1 The time was 00:00:06.6138706 for: iteration 1 Training: 1.8252200355945 The learning rate is 2 ** -12.49694 Starting timing for: iteration 2 The time was 00:00:06.6876942 for: iteration 2 Training: 1.72934456814683 The learning rate is 2 ** -12.49694 Starting timing for: iteration 3 The time was 00:00:06.5961693 for: iteration 3 Training: 1.67938965100877 The learning rate is 2 ** -12.49694 Starting timing for: iteration 4 The time was 00:00:06.5779034 for: iteration 4 Training: 1.65121341726455 ```
No, the initialization does not matter much once KFAC enters the picture which is to be expected. The lack of biases might be an issue.

At any rate this is good.

The next thing on the list would be to make `mapi_concat`. After I do that I will have absolutely everything I need to make my perfected RNNs. Damn, I am really fired up about this.

KFAC is what will make it possible for me to create the perfected plastic RNN. If it has this sort of effect on the vanilla RNN, then it really is plausible that it will let me remove the MI aspects from the one I have now and use modulation for both the outputs and the weight updates.

Let me take a break here.