It does not seem like I will find time to do any programming today as I have some business to take care of. It is just as well. After I've decided on the RNN design yesterday, I now find myself thinking about how to do the generic concat. Right now I have the hastily constructed one that allows concatenating along the inner dimension, but concatenation in general is important enough that I have do deal with it in its full generality.
I said I would do it in the `CudaFun` module, but what I am going to have to do is make `concat_concat` in the `CudaKernel` to implement it. That would be for the best. After that I'll be able to add biases and fuse map and concats with AD inside them. That will be great.

As with every new innovation, dealing with it to a satisfactory degree will take some time.