"5:15pm. I am finally done with today's work. The second part of winter preparations will start from here, but I should be done with it in a week or two.
5:40pm. I know that despite my wishes I did not write a line of code today, but just now I was totally engrossed into the design of the `concat`.

It is not even a `concat` - its true form is a `segmented_init`. It is an upgraded form of `init`. I know that something like this is obvious, but I was not thinking about it like until now.

Those tree views are such a stroke of genius. It is like having them as a piece in the library drives and expands my understanding of both what tensors fundamentally are and how they should be used. Right now, I am feeling that familiar feeling of my power going up. Those experiences are called crystallization. They are the true reward of doing this kind of work.

I'll start work on the `segmented_init` tomorrow in the `CudaKernel` module, make its `CudaFun` relative and then finally bring it into the `Learning` module.

There is a fair bit of work to go through, it will definitely take me at least few days and probably more.

5:55pm. Right now I am tired, so I will call it a day here."