Let me take a break here.
One one hand these new capabilities will be game changing, but on the other hand they are just a bunch of conversion functions that I need to implement. I really do need to think about them for a while. Just now I was deeply engrossed for over an hour. I am starting to get a feel for them.