"10:35pm. https://arxiv.org/abs/1810.08297 Dynamic Automatic Differentiation of GPU Broadcast Kernels
I'll read this paper tomorrow.

10/23/2018

8:45am. I got up 20m ago and have been chilling. Let me start. I want to get the testing done.

But before that, there is something I've been wondering about. I think that residual temporal connections should be an improvement over standard outputs in vanilla RNNs.

The HORNN paper is not quite what I had in mind. They had a lot of stuff in it, but not the very simplest thing which would be linear addition of states at the top level.

Well, it is fine.

Forget about that for now.

Right now I am reading that AD paper.

9:15am. https://www.youtube.com/watch?v=R81pmvTP_Ik
Flux: The Elegant Machine Learning Library

Maybe I will watch this. I guess I should take note of what my competitors are doing.

Let me do this for a bit and then I will start programming."