It is fine.
Going into this, I was prepared for debugging to take another hour, but since I've done the work that I had wanted to I will take a break here.

This segmented iter is a very important function as it will serve as the backbone for many more to come.

---

...Ah, I think I just had a great idea of how I could update the plastic weights in place...

Yeah, I see it. Maybe I could also add explicit unlearning to the mix in place of clipping.

...There is an option I had not considered seriously, but in combination with PRONG it might be realistic. Get rid of normalization entirely.

That would simplify such a vast amount of problems that are now facing me.

Though maybe I simply need to look at it from a different angle. The mutable add's reverse does not necessarily need to be a subtraction.

...Enough of this, let me take a break. These last few paragraphs are just my thought stream regarding reversible learning.

It is going to take some effort to grasp the semantics of doing mutable updates in face of AD, but I will manage it.

One idea that occurs to me now is to ditch the Oja update and take a look at the EGHR paper again. It had that normalization term that was not dependent on the weights.