I am finally back and ready to resume. Only two more days of this and the preparations for the winter will be almost complete. That will free up more time to do programming.
The next thing on the list is `segmented_map.` Slowly but surely, it will be time to start the final steps fo the Spiral's ML library. Once O am done with this, poker won't be a problem. Later I will have to deal with convolutional nets for other things, but now it will absolutely be enough.

I definitely had enough of dealing with ML for one lifetime. I just do not feel like dealing with nonsense anymore until the next break comes around.

I just wanted to be a user of it to begin with. I've been a researcher for way too long.

I want the power of RNNs to become my own.

They are the most powerful of all ML models, and differentiable plasticity (and KFAC) is what is needed to complete them.

I've been thinking about jobs - they are anathema to me, but the only time I would actually consider one would be working on compiler support for fledgling GPU replacements. My current skillset would extremely well suited for that and I would probably do that on my own anyway - for things closely aligned to my interests I would relax my stringent rules on cooperation.

But enough of that. Let me make the segment_map function. It should not be hard compared to what came before.