It works. Yeah, I did good today. I thought I would get to `Learning`, but I guess I will leave that for tomorrow. With this I have a complete solution to the problem of concatenation. I will combine it with kernel AD and get what I want in one fell swoop.
I am really eager to try out the KFACed plastic RNN with modulation on both the plasticity and weight updates. I've been looking towards that for the past few days. Once I finish the vanilla RNN and the LSTM, I really will be able to grasp it.

And once I am at that point, I will thrust into the darkness and challenge reversible learning for the first time.

Once I surmount that the prelude will truly be over. The work on Spiral will have paid off.

I want to use this in the real world. Eventually, I will reach that point.