```     inl segmented_init {dim} init s =         inl out =             open CudaAD             s.CudaFun.segmented_init {dim} <| inl dim ->                 Struct.map2 (inl dim init ->                     inl {out} = init dim >>= succ                     primals out                     ) dim init             |> Tensor.unzip             |> Struct.map (drv s)
        {
        out
        bck=met _ ->
            inl from = adjoints (to_dev_tensor out)
            open CudaAD
            s.CudaKernel.segmented_iter {dim} <| inl dim ->
                Struct.iter2 (inl dim init ->
                    inl {out=to bck} = init dim >>= succ
                    inl {bck=bck'} = link_adjoint_view dim {from to=adjoints to}
                    bck(); bck'()
                    ) dim init
        }
```

This is good. I think this is what I intended with the `segmented_init`.

I am looking at this from various angles and it checks out. The function should be nothing more and nothing less.

Now what do I do about these two?

```
    //inl segmented_mapi f in s =
    //    inl dim = Tensor.assert_broadcastable (primals in)
    //    inl in = to_dev_tensor in
    //    open CudaAD
    //    init {dim} (inl cur -> broadcasting_link dim cur in >>= f cur) s

    //inl segmented_map = mapi << const
```

I am thinking about removing the `segmented_map` from `CudaFun`. It is uncertain what its purpose should be to me. `segmented_init` is intended to be a fusion of concatenation and init operations.

...

Yeah, let me just go for it. There is no point on dwelling on this anymore.

I'll do this and then take a break. When I get back I will start work on the utility functions to make `segmented_init` actually work.