Phew, I am done with the rest.  It is really great to have some extra time in the day.
Let me resume. It is time for the debugging gauntlet. Once I get the vanilla RNN going the LSTM will be a piece of cake.

It is time to sharpen my will and push the change through. The story of the ML library will end and a new beginning will start when I have the optimized plastic RNN. There is only a week left until I am done with it. Let me start my new life...and attain my true power.