``` The learning rate is 2 ** -12.49694 Starting timing for: iteration 0 The time was 00:00:06.3980029 for: iteration 0 Training: 2.23228531497229 The learning rate is 2 ** -12.49694 Starting timing for: iteration 1 The time was 00:00:06.3259406 for: iteration 1 Training: 1.8301734428798 The learning rate is 2 ** -12.49694 Starting timing for: iteration 2 The time was 00:00:06.3145174 for: iteration 2 Training: 1.73374821590123 The learning rate is 2 ** -12.49694 Starting timing for: iteration 3 The time was 00:00:06.4132227 for: iteration 3 Training: 1.6815181401702 The learning rate is 2 ** -12.49694 Starting timing for: iteration 4 The time was 00:00:06.4043558 for: iteration 4 Training: 1.64905046006692 ```
Is there really a point to having biases in at all?

At any rate, it works. I can be sure of it.

I want to take a break, but let me do the LSTM first. It should not be too hard.

Edit: Here is the LSTM. Let me give it a try.