"4:40pm.
```
The learning rate is 2 ** -12.49694
Starting timing for: iteration 0
The time was 00:00:08.6401187 for: iteration 0
Training: 2.56929079830811
The learning rate is 2 ** -12.49694
Starting timing for: iteration 1
The time was 00:00:08.5624390 for: iteration 1
Training: 1.99543742439025
The learning rate is 2 ** -12.49694
Starting timing for: iteration 2
The time was 00:00:08.5482938 for: iteration 2
Training: 1.81852430303115
The learning rate is 2 ** -12.49694
Starting timing for: iteration 3
The time was 00:00:08.6018869 for: iteration 3
Training: 1.71684259057785
The learning rate is 2 ** -12.49694
Starting timing for: iteration 4
The time was 00:00:08.5847738 for: iteration 4
Training: 1.65077870091944
```

Yeah, that's my LSTM. Let me reduce the LR a little.

```
The learning rate is 2 ** -13.00084
Starting timing for: iteration 0
The time was 00:00:08.6345674 for: iteration 0
Training: 2.68430880519479
The learning rate is 2 ** -13.00084
Starting timing for: iteration 1
The time was 00:00:08.6336416 for: iteration 1
Training: 2.11528041020638
The learning rate is 2 ** -13.00084
Starting timing for: iteration 2
The time was 00:00:08.7124078 for: iteration 2
Training: 1.93567445564686
The learning rate is 2 ** -13.00084
Starting timing for: iteration 3
The time was 00:00:08.6509764 for: iteration 3
Training: 1.82449421926182
The learning rate is 2 ** -13.00084
Starting timing for: iteration 4
The time was 00:00:08.6649996 for: iteration 4
Training: 1.74726997223436
```

Nope. Ok, let me increase it then.

4:40pm.

```
The learning rate is 2 ** -12.00083
Starting timing for: iteration 0
The time was 00:00:08.3718330 for: iteration 0
Training: 2.46774691173478
The learning rate is 2 ** -12.00083
Starting timing for: iteration 1
The time was 00:00:08.3502800 for: iteration 1
Training: 1.89262638847782
The learning rate is 2 ** -12.00083
Starting timing for: iteration 2
The time was 00:00:08.3216476 for: iteration 2
Training: 1.72747667663155
The learning rate is 2 ** -12.00083
Starting timing for: iteration 3
The time was 00:00:08.4151908 for: iteration 3
Training: 1.63722404648605
The learning rate is 2 ** -12.00083
Starting timing for: iteration 4
The time was 00:00:08.2841752 for: iteration 4
Training: 1.58000489773557
```

This is pretty decent. It will be even better once I do the mutation optimizations. Let me increase the LR once more. Let's see if can handle it."