``` The learning rate is 2 ** -11.50112 Starting timing for: iteration 0 The time was 00:00:08.3847833 for: iteration 0 Training: 2.45302740866592 The learning rate is 2 ** -11.50112 Starting timing for: iteration 1 The time was 00:00:08.3900558 for: iteration 1 Training: 1.84385565811467 The learning rate is 2 ** -11.50112 Starting timing for: iteration 2 The time was 00:00:08.4291346 for: iteration 2 Training: 1.68252846303478 The learning rate is 2 ** -11.50112 Starting timing for: iteration 3 The time was 00:00:08.3389491 for: iteration 3 Training: 1.60011900962426 The learning rate is 2 ** -11.50112 Starting timing for: iteration 4 The time was 00:00:08.7859162 for: iteration 4 Training: 1.55003905390485 ```
Once more...nope, the error is not improving anymore.

The LSTM is hardly worth using over the vanilla RNN at this point. I am sure that if I jacked up the number of parameters by 4x, the vanilla RNN would be a lot better than the LSTM.

...

Now, I suppose I should move to the plastic RNN, but I am not sure that I feel like it. But I have to.

I have to grit my teeth and go on. I have exactly the amount of steam left to finish just that. It won't be too hard. Everything is set anyway.

Let me do it.