"5:45pm.
```
The learning rate is 2 ** -16
At iteration 9 the cost and the sign accuracy are 22.89144 and 0.942/1.
At iteration 19 the cost and the sign accuracy are 22.8877 and 0.914/1.
At iteration 29 the cost and the sign accuracy are 21.39508 and 0.948/1.
At iteration 39 the cost and the sign accuracy are 20.66674 and 0.93/1.
At iteration 49 the cost and the sign accuracy are 20.5676 and 0.89/1.
At iteration 59 the cost and the sign accuracy are 22.72059 and 0.86/1.
At iteration 69 the cost and the sign accuracy are 21.92247 and 0.85/1.
At iteration 79 the cost and the sign accuracy are 19.99414 and 0.88/1.
At iteration 89 the cost and the sign accuracy are 20.14256 and 0.874/1.
At iteration 99 the cost and the sign accuracy are 19.52369 and 0.888/1.
At iteration 109 the cost and the sign accuracy are 20.15575 and 0.87/1.
At iteration 119 the cost and the sign accuracy are 19.15454 and 0.874/1.
At iteration 129 the cost and the sign accuracy are 20.94604 and 0.846/1.
At iteration 139 the cost and the sign accuracy are 24.1587 and 0.822/1.
At iteration 149 the cost and the sign accuracy are 21.80542 and 0.816/1.
At iteration 159 the cost and the sign accuracy are 18.8582 and 0.866/1.
At iteration 169 the cost and the sign accuracy are 28.2893 and 0.79/1.
At iteration 179 the cost and the sign accuracy are 28.35512 and 0.76/1.
At iteration 189 the cost and the sign accuracy are 26.16179 and 0.802/1.
At iteration 199 the cost and the sign accuracy are 33.40876 and 0.738/1.
```

Let me keep lowering the learning rates. There is something here. 94% in the first iteration is nothing to scoff at.

5:50pm.

```
The learning rate is 2 ** -18
At iteration 9 the cost and the sign accuracy are 23.02542 and 0.946/1.
At iteration 19 the cost and the sign accuracy are 22.87734 and 0.958/1.
At iteration 29 the cost and the sign accuracy are 22.4201 and 0.972/1.
At iteration 39 the cost and the sign accuracy are 22.14837 and 0.968/1.
At iteration 49 the cost and the sign accuracy are 22.23037 and 0.93/1.
At iteration 59 the cost and the sign accuracy are 22.96527 and 0.93/1.
At iteration 69 the cost and the sign accuracy are 22.53942 and 0.942/1.
At iteration 79 the cost and the sign accuracy are 21.46665 and 0.968/1.
At iteration 89 the cost and the sign accuracy are 21.23237 and 0.972/1.
At iteration 99 the cost and the sign accuracy are 21.55291 and 0.958/1.
At iteration 109 the cost and the sign accuracy are 21.61901 and 0.952/1.
At iteration 119 the cost and the sign accuracy are 20.72348 and 0.986/1.
At iteration 129 the cost and the sign accuracy are 21.61065 and 0.946/1.
At iteration 139 the cost and the sign accuracy are 22.3617 and 0.904/1.
At iteration 149 the cost and the sign accuracy are 21.91706 and 0.924/1.
At iteration 159 the cost and the sign accuracy are 21.68748 and 0.9/1.
At iteration 169 the cost and the sign accuracy are 21.38863 and 0.936/1.
At iteration 179 the cost and the sign accuracy are 21.73963 and 0.922/1.
At iteration 189 the cost and the sign accuracy are 21.94557 and 0.898/1.
At iteration 199 the cost and the sign accuracy are 20.58676 and 0.936/1.
```

The sign accuracy is excellent, but the error is atrocious.

5:55pm. No, it is not good. I hoped it would go better than this, but the simplistic architecture I am using is just too weak.

The one I had last time was better. It not just a simple matter of putting in KFAC. I am going to have to work a little to find the key to unlocking the true power of plasticity. I have a few ideas. I am going to try them out tomorrow one by one.

I think I should find something that sticks for learning `theta` using gradient descent."