Today was a good day's work for me. I had the whole day to focus on programming and I did what I wanted to. I have no complaints whatsoever.
Tomorrow the mystery of how to learn the modulation coefficients is waiting for me.

I am aware that perhaps using gradients to do this might not be the best idea, and that have a static learning rate in a range might allow the network to learn longer term dependecies than backprop allows, but for poker, I want it all done using SGD. I won't allow any compromises. I will give learning weight updates using modulation a chance it deserves.

Centralizing the weights for KFAC was not solely for the sake of speed. It was also so I could stabilize the modulatory dynamics in the plastic RNN.

Right now I have everything set to make this work. I will figure out what the issue is here and then move to the final steps for the ML library.