"11:45am. The stupid thing just won't compile.
I give up for now. Right now I am both impatient to have breakfast and annoyed. Better leave that for a bit later when I have time to properly do it.

...Ah, I get it. It is those damn zeros again.

11:55am. KFAC is killing me here. It is killing me! I am not using it on biases and yet it is still not working.

Didn't it work last time on this? I do not get it. I am going to have to do some in depth research.

I am so disappointed in it."