I need to get my bearings in order. The plastic RNN has been one blow after another to me. KFAC was supposed to stabilize things, not make them even worse.
I remember it working last time. Let me backtrack to an earlier commit.

The tree view arc is thankfully coming to a close. For the next few weeks I will dedicate my time to brining the plastic RNN to a close. I need this thing to work. If I could just push a little bit further I could ensure my victory for all the subsequent tasks that are to come.