"2pm. Huh...ok, I understand it now.
My reaction to all of this is what the fuck?!

Do you know why the previous time worked? It is pure luck - the unused call to `expand_singular` pushed up the learning rates for the biases by accident which made the whole thing work even with KFAC. What that did is make the covariance matrices for the biases 0.1 which comes to 10 on the diagonal when they are inverted. This had the effect of making the learning rates for the biases large.

2:15pm. What was I even thinking here?

2:25pm. I am still in thought. I am thinking what could I possibly do.

2:35pm. ...

2:40pm. I need to make up my mind. No matter how I look at it here I have a few choices...

1) Dig deeper into what those pushed up biases are doing. I do not think this is an option for me. I do not have what it takes to analyze this.

2) Ditch KFAC.

3) Change the cost function. It is just unusual how poorly this works. I am coming up with various theories, but the thing that comes out to the forefront of my mind is that maybe KFAC does in fact need a probabilistic cost function and won't work on just anything.

I could accept KFAC not working much better than SGD on this problem, but that it cannot work at all with the plastic RNN I made yesterday is telling me something is deeply wrong.

KFAC should never make things worse.

2:45pm. That last thing is what I need to lock into.

2:50pm. Let me stop here for a bit.

When I get back I'll take a look at the KFGN paper again."