https://arxiv.org/abs/1810.06721 Optimizing Agent Behavior over Long Time Scales by Transporting Value
60 pages. I'll read this later. Let me finally resume.

Yesterday, I put quite a bit of effort in order to get everything in shape, but today it feels like I am worn out before even starting.

I need to come to a conclusion whether KFAC does not work because of the architecture or because of the cost function.

Let me take a look at KFGN paper. I want to see what sort of probabilistic approximation they used to compare KFAC with the KFGN.