"4:25pm. I am dead. I've been defeated. I'll try asking on the ML sub about KFAC, but my work on differentiable plasticity is dead because of this. Between KFAC and DP, I cannot possibly choose DP for the simple reason that RL just works so poorly without KFAC and DP has yet to prove itself.
Let me pose a question.

[D] How I can get KFAC to work with differentiable plasticity?

"Not only does [KFAC](https://arxiv.org/abs/1503.05671) not work with [DP](https://arxiv.org/abs/1804.02464) - it works much worse than standard SGD. The main cause of instability seems to be the back whitening part.

This is a real shock to me right now as I got it to work a week ago, only to figure out today that it was due to a mistake on my part. The updates to the covariance matrices for the biases were done incorrectly which caused them to be updated with zeroes. This had the effect of hiking the learning rates for the biases by 10x which somehow made it work.

I am not sure what to do here and I would not mind some advice. I'll go into more details on the architecture I am using per request.

Now that I've run into this, I am especially interested in hearing what KFAC/PRONG does not work with. Maybe my main issue would be assuming KFAC would be model agnostic and I need to calculate the exact [GN matrix](https://arxiv.org/abs/1706.03662) instead for the backwards step. But the issue with that is that I haven't planned on needing to backpropagate those kinds of block matrices and I wouldn't know how to calculate the derivatives for them whether that be by hand or automatically."

5:35pm. I won't post it as I do not have the mental energy to deal with this anymore. I think I will call it a day here instead. This issue has been boring a hole in my mind for over 3.5 hours now.

I am thinking of what I am going to and I am starting to think that maybe I could apply KFAC just to the standard weights. The modulation weights I should leave alone.

If that fails, then that would mean that KFAC is fundamentally incompatible with plasticity.

5:45pm. Look at it this way, this is an opportunity for me to do some real research. This whole phase will be over by the end of next month anyway and at that point I will make a renewed foray into RL.

I'll give DP a week and then toss it aside if it cannot life up to its potential.

I won't let myself be held back here."